Add serial FreeCAD file passing to 3d geo build
This update also changes the name of the serialization function. We add in the option to pass it directly to the build_3d_geometry so that Docker environments don’t need access to a shared drive. We also left the ability to automatically generate it, and added some sanity checks to avoid errors.